refactor: renames some files and methods
