+++ lib/common/orderbook/presentation/cubit/orderbook_cubit.dart
@@ -2,21 +2,21 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:polkadex/common/orderbook/data/models/orderbook_model.dart';
 import 'package:polkadex/common/orderbook/domain/entities/orderbook_entity.dart';
-import 'package:polkadex/common/orderbook/domain/usecases/fetch_orderbook_data_usecase.dart';
-import 'package:polkadex/common/orderbook/domain/usecases/fetch_orderbook_live_data_usecase.dart';
+import 'package:polkadex/common/orderbook/domain/usecases/get_orderbook_data_usecase.dart';
+import 'package:polkadex/common/orderbook/domain/usecases/get_orderbook_updates_usecase.dart';
 
 part 'orderbook_state.dart';
 
 class OrderbookCubit extends Cubit<OrderbookState> {
   OrderbookCubit({
-    required FetchOrderbookDataUseCase fetchOrderbookDataUseCase,
-    required FetchOrderbookUpdatesUseCase fetchOrderbookUpdatesUseCase,
+    required GetOrderbookDataUseCase fetchOrderbookDataUseCase,
+    required GetOrderbookUpdatesUseCase fetchOrderbookUpdatesUseCase,
   })  : _fetchOrderbookDataUseCase = fetchOrderbookDataUseCase,
         _fetchOrderbookUpdatesUseCase = fetchOrderbookUpdatesUseCase,
         super(OrderbookInitial());
 
-  final FetchOrderbookDataUseCase _fetchOrderbookDataUseCase;
-  final FetchOrderbookUpdatesUseCase _fetchOrderbookUpdatesUseCase;
+  final GetOrderbookDataUseCase _fetchOrderbookDataUseCase;
+  final GetOrderbookUpdatesUseCase _fetchOrderbookUpdatesUseCase;
 
   Future<void> fetchOrderbookData({
     required String leftTokenId,
