+++ test/common/orderbook/domain/fetch_orderbook_live_data_usecase_test.dart
@@ -2,17 +2,17 @@ import 'package:dartz/dartz.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:polkadex/common/orderbook/domain/repositories/iorderbook_repository.dart';
-import 'package:polkadex/common/orderbook/domain/usecases/fetch_orderbook_live_data_usecase.dart';
+import 'package:polkadex/common/orderbook/domain/usecases/get_orderbook_updates_usecase.dart';
 
 class _OrderbookRepositoryMock extends Mock implements IOrderbookRepository {}
 
 void main() {
-  late FetchOrderbookUpdatesUseCase _usecase;
+  late GetOrderbookUpdatesUseCase _usecase;
   late _OrderbookRepositoryMock _repository;
 
   setUp(() {
     _repository = _OrderbookRepositoryMock();
-    _usecase = FetchOrderbookUpdatesUseCase(orderbookRepository: _repository);
+    _usecase = GetOrderbookUpdatesUseCase(orderbookRepository: _repository);
   });
 
   group('FetchOrderbookLiveDataUseCase tests', () {
